requirements.txt
python-dotenv
pytelegrambotapi
python-qbittorrent